Add a button to turn changes on/off for comparison
